Struct hotshot_primitives::vid::advz::Advz

pub struct Advz<P, T, H>where
    P: PolynomialCommitmentScheme,{ /* private fields */ }

Context data for the ADVZ VID scheme.

This struct is a concrete impl for VidScheme. Generic parameters T, H are needed only to express trait bounds in the impl for VidScheme.

H is a hasher.

T is a group.

Implementations

impl<P, T, H> Advz<P, T, H>where P: PolynomialCommitmentScheme,

pub fn new( payload_chunk_size: usize, num_storage_nodes: usize, srs: impl Borrow<P::SRS> ) -> Result<Self, VidError>

Return a new instance of Self.

Errors

Return VidError::Argument if num_storage_nodes < payload_chunk_size.

impl<P, T, H> Advz<P, T, H>where P: PolynomialCommitmentScheme<Point = <P as PolynomialCommitmentScheme>::Evaluation>, P::Polynomial: DenseUVPolynomial<P::Evaluation>, P::Commitment: From<T> + AsRef<T>, T: AffineRepr<ScalarField = P::Evaluation>, H: Digest + DynDigest + Default + Clone + Write,

pub fn dispersal_data_from_elems( &self, payload: &[P::Evaluation] ) -> Result<(Vec<<Self as VidScheme>::StorageShare>, <Self as VidScheme>::StorageCommon), VidError>

Same as VidScheme::dispersal_data except payload is a slice of field elements.

pub fn recover_elems( &self, shares: &[<Self as VidScheme>::StorageShare], _common: &<Self as VidScheme>::StorageCommon ) -> Result<Vec<P::Evaluation>, VidError>

Same as VidScheme::recover_payload except returns a Vec of field elements.